feat : 신규 도메인(엔터티) 추가 - 기존 엔터티는 legacy 폴더 하위로 이동
