+++ src/main/java/com/musinsa/demo/legacy/domain/User.java
@@ -1,0 +1,20 @@
+package com.musinsa.demo.legacy.domain;
+
+import lombok.*;
+
+import javax.persistence.*;
+
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class User {
+    @Id
+    @Column(name = "user_no")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long userNo;
+    @Column(unique = true, name = "user_id")
+    private String id;
+}
